pyproject.toml
[project]
name = "namecheap-email-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.131.0",
    "fastapi-mcp>=0.4.0",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
